ERM-3580/ERM-3581 Skip ingest of titles for packages where syncContentsFromSource == false (#857)
* env: Add .env to gitignore

* feat: Prevent syncing of titles via EITHER harvest or pushKB when syncContentsFromSource is set to false.

In addition, set GOKbOAIAdapter to automatically set packages to `syncContentsFromSource:  false`.

To be confirmed -- this should NOT overwrite `syncContentsFromSource:  null`

* test: Added syncContentsFromSource to all test JSON

* feat: SYNC_PACKAGES_VIA_HARVEST

Added env var SYNC_PACKAGES_VIA_HARVEST which allows the turning on/off of the default behaviour to sync packages coming in via harvest process

Also added readme section for env vars and filled in this row

* test: PackageSyncSpec

Replaced PackageImportSpec with a slightly more involved PackageSyncSpec. Include fixes for WorkSourceIdentifierTIRSIngestJobSpec with/without the env variable in place

Also attempt to fix the github actions so that WorkSourceIdentifierTIRSIngestJobSpec runs as before and therefore checks everything required of it

* ci: swap syncViaHarvestEnvVar default in test runner

Set syncViaHarvestEnvVar to FALSE by default instead of true

* test: Fix test failure in PackageSyncSpec

Package names were not lining up properly

* test: Add case to check syncStatus does not get overwritten

* test: Add case to check syncStatus can't be changed directly via PUT (Need to check cases where Pkg is included in some other API call such as agreementLine)

* feat: /erm/packages/controlSync

New endpoint to allow users to control sync status on multiple packages

WIP! Does not currently do ANY work to send a request back to pushKB, or utilise any metadata on the package, and it is ALSO not currently exposed through Okapi

refs ERM-3581, ERM-3580

* build: Applied controlSync endpoint to Okapi module descriptor

Also changed from PUT to POST in line with rest of module

* test: Add test cases for controlSync working as expected

* docs: Added some more of the env variables to README