ERM-3438: Agreement start date not showing as updated after period updated (#880)
* test: Update start date integration test

Created integration test to demonstrate the issue in which the start date is not updated when the agreement has a custom property with the type of "Multi-select"

* refactor: Rename spec to reference specific issue

When we have hyper specific test cases for issues I see no problem with referencing the issue number in the spec name.

It makes it very wordy but tbh that's fine, mostly you'd just run the tests.

* refactor: Another tweak to spec name, I changed my mind slightly

* test: Added test cases. It's now clear that bug occurs when _any_ multi value refdata is in the system, whether or not that is the one we're binding to the agreement

* test: Oops, previous commit wasn't the case. Have changed test wording round to make that clear

* test: Added integer custprop setup and test cases to see if it fails on all multi-type custprops

* fix: Initial attempt at a fix

Added explicit "markDirty" when transiently bound fields are changed in SubscriptionAgreement.

I do not yet understand why the dirty checks fail in the multi-refdata case and not elsewhere, but this seems to fix the issue at hand.

I'd like to come back to this and fully understand the behaviour

* chore: Updated comment

---------

Co-authored-by: Jack Golding <jack.golding@k-int.com>