fix: Keep "internal" entitlements with type = null for now -- ERM-3795 (#946)
Internal entitlements historically had a null type. The comments elsewhere in the code suggested that `internal` was an acceptable value, but this causes frontend issues

For now, simply revert to using `null` here for our logic, but eventually this needs a proper rethink. String value for something logic or rendering keys off is a terrible idea

refs ERM-3795